moved parser to same folder as notebooks. created medium notebook. amended coarse to use parser class
